MIC-02 refactor: Feature/implementa repositories (#13)
* MIC-2 refactor: altera tipo dos ids nos repositories para Long

+++ src/main/java/net/burnbook/app/repository/ComentarioRepository.java
@@ -12,7 +12,7 @@ import java.awt.print.Pageable;
 import java.util.List;
 
 @Repository
-public interface ComentarioRepository extends JpaRepository <Comentario, Integer> {
+public interface ComentarioRepository extends JpaRepository <Comentario, Long> {
 
     Page<Comentario>
     findByPublicacaoAndComentarioPaiIsNullOrderByDataHoraDesc(Publicacao publicacao, Pageable pageable);
